// types/student.ts
export interface StudentProfile {
    id?: string
    first_name?: string
    last_name?: string
    university?: string
    major?: string
    graduation_year?: number
    skills?: string[]
    bio?: string
    avatar_url?: string
    resume_url?: string
    created_at?: string
    updated_at?: string
  }